Refactor /data. Update README.
