+++ .gitignore
@@ -7,3 +7,6 @@ __pycache__/
 /*.egg-info/
 *.egg-info/
 *.egg
+
+#Individual files
+pyproject_deprecated.toml
