+++ cytopus/KnowledgeBase/__init__.py
@@ -1,5 +1,5 @@
 """Tools to plot and query KnowledgeBase"""
-from .kb_queries import KnowledgeBase
+from .kb_queries import KnowledgeBase, get_data
