--- pyproject.toml
@@ -1,26 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "cytopus"
-version = "1.3.0"
-authors = [
-  { name="Thomas Walle", email="t.walle@dkfz.de" },
-]
-description = "A Knowledge Base for Single Cell Biology"
-readme = "README.md"
-requires-python = ">=3.7"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
-]
-dependencies = [
-	"pandas>1.3",
-	"numpy>1.2",
-	"networkx>2.7",
-	"matplotlib>3.4"]
-[project.urls]
-"Homepage" = "https://github.com/wallet-maker/cytopus"
-"Bug Tracker" = "https://github.com/wallet-maker/cytopus/issues"
